+++ RELEASES.md
@@ -4,7 +4,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * module-tools: prov-template-archetype and prov-toolbox-plugin
 * prov-template-compiler: record index with principal
 * service-template: simple and composite queries with access control
-
+* service-template: queries with access control, incl browsing and navigation
 
 `2.0.7`
 * prov-jsonld: fixed context url